requirements.txt
fastapi[standard]
uvicorn
python-dotenv
supabase
python-multipart
numpy
huggingface-hub
sentence-transformers
transformers
openai
vecs
langchain
langchain_community 
langchain_text_splitters
pypdf
openai
sqlalchemy
alembic
pydantic
pydantic-settings
asyncpg
pgvector
greenlet
psycopg2